updated dev workflow
Added auto-reload functionality on file change

Origially, the firebase server had to be manually restarted with every file change. To circumvent this, Nodemon was installed to restart the server when the compiled JS changed. Another script had to be created to serve the functions without re-building the app, as the original built the app and then served. This would trigger infinate loops where the changes in the lib directory would rebuild the application, thus retriggering the build. Now there is a script to serve the application without rebuilding.