Restructured Transaction Service
Restructured the transactions service so that it can fetch not only incomes, but also expenditures, as well as being able to fetch both together. I also added the property to the model that determines the type of transaction veing operated on. There is also now a set of functiona that can be used to map the data for delivery to the client.